Rename handler modules to drop _handler.
